Move "testMergedDocumentsUsingPrecursorLevelDataScenario" into separate class so that it can use a different project name. Refactor functionality from TestMSCalibrationCurveTest into "AbstractQuantificationTest" so that it can be shared with "TargetedMSPrecursorLevelDataTest".
